libs/inference/pyproject.toml
[project]
name = "inference"
version = "0.1.0"
description = "Shared inference library for model loading and agent factories"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "shared",
    "langchain>=0.1.0",
    "langgraph>=0.0.10",
    "jinja2>=3.1.0",
    "tenacity>=8.2.0",
    "pydantic>=2.0.0",
    "google-cloud-aiplatform>=1.38.0",
    "requests>=2.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/inference"]

[tool.ruff]
lint.select = ["E", "F", "W", "C", "N", "I", "D"]
lint.ignore = ["E203", "D203", "D213", "D100", "D413", "D104"]
lint.pydocstyle.convention = "google"
target-version = "py312"
line-length = 88
exclude = ["venv", ".venv", "docs"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "docs"]
plugins = []
mypy_path = ["../shared/src"]

[tool.pytest.ini_options]
pythonpath = ["src", "../shared/src"]
testpaths = ["tests"]
addopts = "--cov=inference --cov-branch --cov-report=term-missing"

[tool.uv.sources]
shared = { workspace = true }
